多类型表：优化一点细节问题
